requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
pyjwt==2.8.0
pytz==2023.3.post1